# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EzCon API",
        "description": "API to manage condominous, \r\n to get the authorization working, put `Bearer your-token-in-authorize-field`, \r\n whitout the *Bearer* you will not be auhtorized to use the services",
        "contact": {
            "email": "eikesantz@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://cors-anywhere.herokuapp.com/https://ezcon-api.herokuapp.com/",
            "description": "API hosted on Heroku (passing by proxy)"
        },
        {
            "url": "http://ezcon-api.local/",
            "description": "API local"
        }
    ],
    "paths": {
        "/users/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get logged user information",
                "description": "Returns user's iformotation",
                "operationId": "getUsersList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/roles": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Add role to a user",
                "description": "Returns user with roles",
                "operationId": "addUserRoles",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Remove role from user",
                "description": "Returns user with roles",
                "operationId": "removeUserRoles",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/signIn": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User's login",
                "description": "Make user's login and retrive the access token",
                "operationId": "loginUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "user@email.com",
                                    "password": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "402": {
                        "description": "Missing some parameter"
                    }
                }
            }
        },
        "/signUp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User's registration",
                "description": "Create a new user",
                "operationId": "registerUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirmPassword": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "user",
                                    "email": "user@email.com",
                                    "password": "123456",
                                    "confirmPassword": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "402": {
                        "description": "Missing some parameter"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Apartment": {
                "title": "Apartment",
                "description": "Apartment model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Apartment's id",
                        "type": "integer",
                        "example": "1"
                    },
                    "number": {
                        "title": "number",
                        "description": "Apartment's number",
                        "type": "integer",
                        "example": "104"
                    },
                    "block": {
                        "title": "block",
                        "description": "Apartment's block",
                        "type": "string",
                        "example": "A"
                    }
                },
                "type": "object"
            },
            "Role": {
                "title": "Role",
                "description": "Role model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Role's id",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "Role's name",
                        "type": "string",
                        "example": "admin"
                    },
                    "description": {
                        "title": "description",
                        "description": "Role's description",
                        "type": "string",
                        "example": "Administrator can do everything"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "User's id",
                        "type": "integer",
                        "example": "8"
                    },
                    "name": {
                        "title": "name",
                        "description": "User's name",
                        "type": "string",
                        "example": "user"
                    },
                    "email": {
                        "title": "email",
                        "description": "User's email",
                        "type": "string",
                        "example": "user@email.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "User's password",
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Use the Auth routes and recieve your accessToken",
                "name": "Authorization",
                "in": "header",
                "scheme": "https"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Operations to register and authorize the user"
        },
        {
            "name": "User",
            "description": "Operations about user"
        }
    ]
}